+++ gazebo/transport/Publisher.cc
@@ -192,10 +192,12 @@ void Publisher::SendMessage()
         iter != localBuffer.end(); ++iter, ++pubIter)
     {
       // Send the latest message.
-      this->pubIds[*pubIter] = this->publication->Publish(*iter,
+      int result = this->publication->Publish(*iter,
           boost::bind(&Publisher::OnPublishComplete, this, _1), *pubIter);
 
-      if (this->pubIds[*pubIter] <= 0)
+      if (result > 0)
+        this->pubIds[*pubIter] = result;
+      else
         this->pubIds.erase(*pubIter);
     }